moved components around
