--- client/components/ItemDetails.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export default function ItemDetails() {
-  return (
-    <>
-      <div className="itemdetails"></div>
-    </>
-  );
-}
